partypad-backend/requirements.txt
alembic==1.8.1
asgiref==3.5.2
boto3==1.24.91
botocore==1.27.91
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
direnv==2020.12.3
Django==4.1.1
django-cors-headers==3.13.0
django-extensions==3.2.1
django-ses==3.2.0
djangorestframework==3.14.0
docopt==0.6.2
env-file==2020.12.3
greenlet==1.1.3
idna==3.4
jmespath==1.0.1
Mako==1.2.3
MarkupSafe==2.1.1
numpy==1.23.4
pandas==1.5.1
Pillow==9.3.0
pip==22.2.2
pipreqs==0.4.11
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
pyOpenSSL==22.1.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
s3transfer==0.6.0
setuptools==63.4.1
six==1.16.0
SQLAlchemy==1.4.41
sqlparse==0.4.3
urllib3==1.26.12
values==2020.12.3
Werkzeug==2.2.2
wheel==0.37.1
yarg==0.1.9
